#9 refactor: フォームの内容を全て別のコンポーネントに置き換える
